rename v2-footer to footer, replace original footer
