--- blocks/v2-footer/v2-footer.css
@@ -1,245 +1,0 @@
-:root {
-  --border-color: var(--c-grey-300);
-}
-
-.v2-footer {
-  background: var(--c-grey-50);
-}
-
-.v2-footer :is(.v2-footer__title, p, a) {
-  font-family: var(--ff-volvo-novum);
-  font-feature-settings: 'clig' off, 'liga' off;
-  font-size: 16px;
-  letter-spacing: 0.5px;
-}
-
-.v2-footer ul {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.v2-footer-links-wrapper :is(.v2-footer__title, li, p, a) {
-  margin-bottom: 0;
-  color: var(--c-main-black);
-}
-
-.v2-footer-list .v2-footer__title {
-  font-family: var(--ff-volvo-novum-medium);
-  border-top: 1px solid var(--border-color);
-  cursor: pointer;
-  padding: 28px 16px;
-  display: flex;
-  align-items: center;
-}
-
-.v2-footer-list-item {
-  transition: 300ms max-height ease;
-  will-change: max-height;
-  overflow: hidden;
-  max-height: 0;
-}
-
-.v2-footer-list-item a {
-  display: flex;
-  padding-left: 16px;
-  padding-right: 16px;
-  padding-bottom: 16px;
-}
-
-.v2-footer-list .icon-chevron-down {
-  margin-left: auto;
-  transition: transform var(--duration-medium) cubic-bezier(0.45, 0, 0.125, 1);
-}
-
-.v2-footer-list.expand .icon-chevron-down {
-  transform: rotate(-180deg);
-}
-
-.v2-footer-list .icon-newtab,
-.v2-footer-list .icon-fb,
-.v2-footer-list .icon-twitter,
-.v2-footer-list .icon-yt,
-.v2-footer-list .icon-instagram,
-.v2-footer-list .icon-linkedin {
-  display: none;
-}
-
-.v2-footer-copyright-wrapper {
-  border-top: 1px solid var(--border-color);
-}
-
-.v2-footer-copyright-container {
-  display: flex;
-  padding: 12px 16px;
-  flex-flow: column;
-  align-items: center;
-  gap: 6px;
-  flex-wrap: wrap;
-}
-
-.v2-footer-copyright {
-  text-align: center;
-  flex-basis: 100%;
-}
-
-.v2-footer-bar {
-  display: flex;
-  padding-top: 12px;
-  padding-bottom: 12px;
-  justify-content: flex-end;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-.v2-footer-bar li {
-  flex: 1 0 auto;
-}
-
-.v2-footer-bar a,
-.v2-footer-copyright p {
-  color: var(--c-grey-4);
-  margin: 0;
-  padding-top: 8px;
-  padding-bottom: 8px;
-}
-
-.v2-footer a {
-  text-decoration: none;
-}
-
-.v2-footer a:hover {
-  color: var(--c-cta-blue-hover);
-  text-decoration: underline;
-}
-
-.v2-footer-bar li:not(:first-child) {
-  padding-left: 8px;
-}
-
-.v2-footer-bar li:not(:last-child) {
-  border-right: 1px solid var(--volvo-text-gray);
-  padding-right: 8px;
-}
-
-.v2-scroll-to-top {
-  background-color: var(--c-white);
-  border-radius: 50%;
-  border: 1px solid var(--c-grey-400);
-  bottom: 20px;
-  box-shadow: 0 4px 70px 0 rgb(0 0 0 / 20%);
-  color: var(--c-main-black);
-  cursor: pointer;
-  display: none;
-  padding: 10px;
-  position: fixed;
-  right: 30px;
-  z-index: 99;
-}
-
-.v2-scroll-to-top:hover {
-  background-color: var(--c-grey-50);
-}
-
-.v2-scroll-to-top:active {
-  background-color: var(--c-grey-100);
-}
-
-.v2-footer-list:not(.expand) .v2-footer__title:focus,
-.v2-footer-list-item a:focus,
-.v2-footer-bar a:focus,
-.v2-scroll-to-top:focus {
-  outline: 2px solid var(--light-border-focus);
-}
-
-.v2-scroll-to-top svg {
-  display: block;
-}
-
-/* desktop */
-@media screen and (min-width: 1200px) {
-  .v2-footer-links-wrapper,
-  .v2-footer-copyright-container {
-    width: calc(var(--wrapper-width) - 100px);
-    margin: 0 auto;
-  }
-
-  .v2-footer-links-wrapper {
-    padding-top: 55px;
-    padding-bottom: 55px;
-  }
-
-  .v2-footer-list .v2-footer__title {
-    border-top: 0;
-    padding: 0;
-    cursor: initial;
-  }
-
-  .v2-footer-list:not(.expand) .v2-footer__title:hover {
-    background-color: transparent;
-  }
-
-  .v2-footer-list-wrapper {
-    display: flex;
-    max-width: 1200px;
-    gap: 20px
-  }
-
-  .v2-footer-list, 
-  .v2-footer-list-item {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .v2-footer-list {
-    flex: 1;
-    gap: 24px;
-  }
-
-  .v2-footer-list-item {
-    gap: 20px;
-    max-height: 100%;
-    overflow: auto;
-  }
-  
-  .v2-footer-list-item a {
-    gap: 16px;
-    align-items: center;
-    padding: 0;
-  }
-
-  .v2-footer-list .icon-chevron-down {
-    display: none;
-  }
-
-  .v2-footer-list .icon-newtab,
-  .v2-footer-list .icon-fb,
-  .v2-footer-list .icon-twitter,
-  .v2-footer-list .icon-yt,
-  .v2-footer-list .icon-instagram,
-  .v2-footer-list .icon-linkedin {
-    display: flex;
-  }
-
-  .v2-footer-copyright-container {
-    flex-direction: row-reverse;
-    justify-content: space-between;
-    align-items: center;
-    padding-left: 0;
-    padding-right: 0;
-  }
-
-  .v2-footer-bar {
-    flex-flow: row;
-  }
-
-  .v2-footer-copyright {
-    flex-basis: auto;
-  }
-
-  .v2-footer-text-icon-wrapper {
-    display: flex;
-    gap: 4px;
-  }
-}
